backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
transformers==4.35.2
torch==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
email-validator==2.1.0
beautifulsoup4==4.12.2
requests==2.31.0

